Allow 30 byte empty buffer during connection setup
A several second delay is observed when using the O2 mobile network to performa TWAMP test,
this delay is in the TCP setup, as the client waits for the server to send a greeting something in
the network is delaying the connection to the server. A workaround for this is to send some data from
the client to the server, this immediately opens the connection to the server.